I added the norm of the constraint to the output.
I added the norm of the constraint to the output.  This makes more
visible a problem that I've been working with.  Typically, we use the
letter g for the gradient, be it the unconstrained gradient or the
gradient of the Lagrangian.  However, the equality constraint is also
labeled by g.  In the code, we can prevent the conflit since the state
stores the gradient and the bundle of functions stores the equality
constraint, but it is a little irritating and should be fixed.