First, I added the complete output for the...
First, I added the complete output for the equality constrained code,
which means that I added the Krylov method information.  Second, I
cleaned up the variable names, which changes slightly the API for the
users.  Specifically  1.  eps_g -> eps_grad -- This is necessary because
we have g as a function and also as a gradient.  In order to avoid this
conflict, g became grad. 2.  norm_g -> norm_grad 3.  norm_gtyp ->
norm_gradtyp 4.  g -> grad 5.  g_old -> grad_old 6.  External ->
UserDefined -- Originally, the name External was used due to another
project.  However, frankly, it's a little confusing since External means
that it's external to some sort of code, but it's not clear which one.
UserDefined makes it much more clear that we're using the UserDefined
Hessian. 7.  Minv_type -> PH -- Many references in Krylov methods and
optimization like to use M^{-1} for the preconditioner, which is why we
did so originally.  There's not wrong with this, but I started getting
tired of carrying around and inverse.  Honestly, we never have the
forward operator.  Ever.  As such, it doesn't make a lot of sense to say
that we have an inverse when, really, we just have some sort of operator
that we use as a preconditioner.  In addition, the equality constrained
code requires two additional preconditioners for the augmented system,
so it becomes less clear what we're preconditioning.  As such, the new
scheme is that capital P means preconditioner and the H means Hessian.
Hence, PH means we have a preconditioner for the Hessian. 7.  We'll, it
really was Minv_type -> PH_type, but I can't type (pun?)   8.  Minv ->
PH 9.  Minv_left_type -> PSchur_left_type 10. Minv_right_type ->
PSchur_right_type 11. Minv_left -> PSchur_left 12. Minv_right ->
PSchur_right 13. TR_op -> TRS -- I was kind of irritated that I kept
have to type _op.  Basically, I try to use capital letters when I have
an operator, so the _op was redundant.  TRS now stands for Trust Region
Scaling or Trust Region Shape. 14. W_gpHdxn -> W_gradpHdxn -- Again,
this is part of the change from g to grad. 15. g_orig -> grad_orig 16.
g_schur -> grad_schur 17. g_lag -> grad_lag  In addition to cleaning up
the code, I modified all of the examples and unit tests to use this new
API.