I fixed a number of bugs related to the...
I fixed a number of bugs related to the tangential subproblem.  Probably
the biggest issue has come from the projection inside of the subproblem.
Basically, in the nw_sqp_exercise problem, we had a problem where the
residual in truncated-CG in the second iteration was in the nullspace of
the constraints when we start with a Lagrange multiplier of [1,1,1].
When this happens, it becomes impossible to meet the inexact stopping
criteria and the method continues.  If it continues for too many
iterations, we get a NaN.  In order to combat this, I've added some
logic to the projection inside of truncated-CG as well as the gradient
projection ahead of time to bail early if this can be detected.  In
order to detect it, I wait until the second iteration of the Krylov
method and then see if the current projected solution is small with
respect to the RHS.  Hopefully, this only happens if we run into this
issue.  Now, I did not modify the stopping conditions for the tangential
step, which may run into this issue as well.  Basically, anytime we
solve an augmented system where one of the stopping criteria is based on
the norm of the current iterate, this could be a problem.  When the
solution is near zero, the stopping criteria becomes near zero plus
whatever the xi tolerance adds, so we get a race to see what becomes
zero first.  In the tangential step computation, we have a norm of ||
dx_n + dx_t || where we're iterating on dx_t.  Hence, this could be an
issue in the future, so we need to be mindful of it.  Next, I needed to
modify truncated CG in order to stop on the norm of the preconditioned
residual.  This stops the method from continuing in the case we're using
a degenerate preconditioner like a projection.  Basically, once the
iterations exceed the rank of the projection, we need to stop and
looking at the norm of the preconditioned residual is the easiest way to
see that because it becomes zero.  I also needed to modify truncated
MINRES.  In the case the preconditioner has rank-1, we were not
calculating the step or the error correctly.  Since we only have one
Krylov vector at this point, we can calculate the step by hand.  Now,
unlike CG, MINRES looks at the preconditioned-norm of the residual, not
the norm of the preconditioned residual.  Since the preconditioner is a
projection in the equality-constrained algorithm and this operator can
be degenerate and/or nonsymmetric, this can and does ocassionally cause
issues.  Basically, we can get nans because things like x'Bx can be
slightly negative when it's near zero.  Right now, the method is doing
an ok job of bailing out in these situations, but it may slightly affect
performance, so it'd be nice to figure out a workaround for this issue.
Anyway, at this point the Nocedal and Wright example is solving in 5
iterations, which seems good.  I may want to add some code to find an
initial Lagrange multiplier, but I don't think that's strictly
necessary.