I added the pieces to run the equality...
I added the pieces to run the equality constrianed code in MATLAB.
Everything seems to work, but I'm noticing some odd delays.  In
addition, there was a bug in peopt where it was calculating the initial
step twice.  This has been eliminated.  I also fixed a problem where the
rejected steps where not creating output.  This was due to the ordering
of how the state manipulators were wrapped.  We should be mindful of
these nuances going forward.