I fixed how the penalty parameter for the...
I fixed how the penalty parameter for the augmented Lagrangian is
calculated and used.  Basically, we weren't keeping track of the old
penalty parameter and constantly updating the parameter in the inner
loop.  This is incorrect.  The penalty parameter should only be updated
between major optimization iterations.  This update fixes that.