In order to reduce the time spent in factoring...
In order to reduce the time spent in factoring something like a matrix
in a precondition, I now cache the result of g'(x)dx_t.  This quantity
was required in the residual predicted reduction.  Basically, any time
we evaluate at a new point, x+dx, we may have to factorize our
preconditioner again.  Hence, it's important not to evaluate anything at
a new iterate until we've exhausted our old.  Now, in the case where we
need to refine our tangential step or reject a trust-region step, we're
going to have to reevaluate functions at x and then at a new x+dx.
Hence, there may still be some benefit in more advanced caching schemes.
Second, there was a bug with how the preconditioner was applied in the
computation of the Lagrange multiplier step.  Basically, for things like
preconditioners, we don't have anything like an operator with a base.
Rather, we just have a simple operator.  This causes problems when the
operator requires knowledge of the iterate in order to evaluate, which
is the case in the Schur preconditioner.  Anyway, the way we get around
this is to create a reference to the current iterate when constructing
the operator.  Now, in the computation of the Lagrange multiplier step,
we need to evaluate the operator at a point x+dx, which means we need
our preconditioner at x+dx.  In order to fake this, we take a step,
solve the Lagrange multiplier step, and then step back.  This is sort of
a dirty hack, so it may be worthwhile in the future to have both an
operator as well as an operator with a base.