I reworked truncated-CG so that we now better...
I reworked truncated-CG so that we now better detect instability.  In
theory, the residual r and the projected residual B r should be
orthogonal.  In practice, with inexactness, this may not be the case.
As such, I implemented a heuristic used by Denis and Matthias where we
construct a matrix inv(D) ( Brs' Br ) inv(D) - I where D=diag(||Wr_i||).
In exact arithmetic, this matrix should have norm 0.  In inexact
arithmetic, it does not.  So, we exit if the Frobenius norm of this
matrix ever grows to be more than 0.5.  In addition, when we detect
instability, we no longer scale the current direction to the
trust-region.  Rather, we just cut off the computation.  These two
changes have dramatically improved the speed and reliability of the
composite step algorithm.