I fixed the compatibility issues with the...
I fixed the compatibility issues with the state manipulators in the
constrained code.  In theory, the fully constrained code should be
working, but I've not tested it.  By the way, there is an interesting
and difficult problem associated with the design of the state
manipulators.  Right now, the state manipulators have to do some dynamic
casting in order to get the state and functions into the correct space.
Specifically, in each of the codes, we hook into the unconstrained
optimization routines.  This is nice that we've been able to do this
convergence because it means that we've reduced the amount of code we
have to write and maintain.  The drawback to this is that the
unconstrained code is designed to work on unconstrained states,
functions, and state manipulators.  Since we use the state manipulators
to add in pieces such as the composite step method and the interior
point method, this is difficult.  So, we call the state manipulator,
convert the state back into either an equality constrained or an
inequality constrained state and do the right thing.  In theory, the
visitor patter would be a good choice for this design since we could use
the double dispatch system of C++ in order to do the safe conversions
for us.  The problem is that our states, functions, really everything is
templated on our real type as well as our vector space types.  This
intereres with our ability to use the visitor pattern and use dynamic
polymorphism to resolve these spaces.  Now, the advantage of this
templating was that we can guarantee we have compatible mathematical
spaces, which has been a huge payoff, but this is an instance where it's
been a pain.  In theory, there's no reason why we can't have both, but
we're pushing into this difficult area where C++ is awkward to deal
with.  In any case, it is worthwhile looking into further whether or not
there's a better design for the state manipulators and code insertion.