I added a fix to the quasinormal step.
I added a fix to the quasinormal step.  Basically, it's possible that
the Cauchy point for the quasinormal step brings us to optimality.  In
this case, the stopping criteria for the Newton step is zero, and we can
never satisfy this condition.  Therefore, we add some early exit
criteria.  Specifically, we have that g(x+dx) ~= g(x)+g'(x)dx.  Hence,
if || g(x)+g'(x)dx || is small we should be feasible after the step.
Note, the quantity || g(x)+g'(x)dx || is used in the stopping critiera
for the quasinormal step.  Now, our stopping condition for optimality
just states that || g(x) || < eps_constr || g(x_typ) ||.  Therefore,
since we really care about our stopping conditions, we can check if our
Cauchy point is going to satisfy the stopping condition for feasibility.
If it does, we just bail on the Newton step.  This seems to be mostly
scale resistant, or at least it puts the onus of what is feasible and
what is not feasible on the user where it belongs.  In addition, I added
a very basic example that does a combined inequality and equality
constrained problem.  It does use the code described above on bailing
early.