I included the code to run MATLAB with fully...
I included the code to run MATLAB with fully constrained problems.  In
addition, I fixed a bug where the trust-region scaling was not taken
into account during the quasinormal step.  This was causing the
tangential subproblem to bail immediately since the shifted trust-region
did not include our initial zero guess.