I reworked the equality constrained code to...
I reworked the equality constrained code to work more properly with the
inequality constrained code.  First, the wrapping of the manipulators
was incorrect.  Basically, getStep is called from inside of the equality
constarined manipulator and getStep should be passed a manipulator.
However, it's hard to pass itself as a manipulator.  Because of this, we
need the interior point manipulator to be on the inside of the equality
constrained manipulator.  In addition, we required better coordination
around who does what each point in time.  For example, in the interior
point method, we restore the non interior point modified gradient prior
to the quasi-Newton update.  However, the composite step code may need
to update the Lagrange multiplier, so it's important to make sure that
gradient of the schur complement system is loaded.  As such, we've a
little more granularity with the state manipulator locations.  Next,
there's this ongoing issue with the Lagrange multiplier.  Basically, the
Lagrange multiplier is generally assumed to minimize the quantity ||
grad f(x) + g'(x)*y || for y.  However, with an interior point method,
grad f(x) actually contains interior point pieces and these pieces could
change after we calculate the lagrange multiplier update.  As such,
there's an open question if we need to update the Lagrange multiplier
when this happens.  The global convergence of the composite step
algorithm doesn't require anything more than boundedness, but that
doesn't mean fast convergence.  This should investigated further.  Next,
the trust-region scaling was stopping the inexact/interior point from
converging properly.  I'm not sure why, but we were taking absurdly
small steps.  As such, I've disabled this code.  In terms of
performance, we can set all of the nominal tolerances very low except
for xi_proj.  This seems to need to remain around 1e-3.  I'm not sure
why this is exactly, but may be related to the illconditioning of the
Hessian.  I was able work around the problem with this tolerance by
setting sigma, the centering parameter, so something really high.
Finally, there's another open question as to whether or not we should
try to update things like the predicted reduction or penalty parameter
when the positivity line-search truncates the step.  The caching for
this is a bit tricker and I'm not sure of the overall computational
expense, but it's something to look into.