I modified the interior point method to make...
I modified the interior point method to make it easier to isolate how we
choose the interior point parameter.  Previously, I implemented a scheme
where we always chose sigma*mu where sigma was the reduction parameter
and mu was the interior point estimate.  This wasn't all that flexible
when we wanted to do more exotic things like limit how low we want the
interior point parameters to go.  As such, the current scheme is to keep
the interior point estimate in the variable mu_est and then consolodate
the interior point parameter into mu.  Then, the simple reduction scheme
is to just allow mu = sigma*mu_est.  Along these lines, I also
implemented a scheme to cap the interior point parameter once we
satisify the stopping condition for complementary slackness.  It's my
hope that this will help mediate the affect of ill-conditioning on the
Hessian and it appears to be having this affect.  Also, I added more
output to the diagnostics.  Now, we also output mu and mu_est for the
interior point method and the actual vs. predicted reduction for
trust-region methods.  Finally, the changes to the interior point method
required a simple rebaseline of a few tests.  This appears to be related
to capping out the interior point parameter, so it should be expected.