This is a pretty significant rework of how the...
This is a pretty significant rework of how the functions are modified
when injecting inequality or equality constrained code into pieces of
the unconstrained algorithm.  Previously, we had our scalar valued
function and then modified it depending on our problem context.
Generally, the evaluation of this function meant we were evaluating the
merit function.  The gradient meant the gradient of the Lagrangian.  The
Hessian-vector product meant the Hess-vector product of the gradient.
Where this became tricky was in the inequality constrained code since we
actually had three different gradients depending on the context.
There's the true gradient of the Lagrangian that we use in our stopping
conditions, grad L(x,y) - h'(x)*z.  There's the Schur complement
gradient, grad L(x,y) - mu h'(x)* (inv(L(h(x))) e), that we use when
finding our trial step (truncated-CG and the like).  There's the
original gradient of the Lagrangian that we use for our quasi-Newton
methods, grad L(x,y).  In order to use each of these gradients, we
cached all of them when we did a computation and then tried to do a
juggling act with the state manipulator, which basically meant that we
copied in the appropriate gradient before the step that required it.
Unfortunately, when introducing both equality and inequalities this
became complicated since there were some ordering issues.  Namely, do we
need to update the interior point gradient and then do a Lagrange
multiplier computation?  Is the current stored gradient the one that
corresponds to the correct interior point parameter?  Technically, we
could have solved these issues with our old design, but it was becoming
increasingly hard to keep track of all of this.  In the new strategy,
all of this becomes explicit.  Now, we only store the true objective
function and gradient of the current iterate.  Then, whenever we need
these other additional gradients, we compute them on the fly.  There's
some additional cost in doing this, but I added some caching, so
hopefully that cost is minimal.  The benefit was that all of the
gradient juggling code disappeared and the problem appears to be more
consistent.  In addition, I added a new interior point reduction
strategy called StairStep.  Basically, it only reduces the interior
point parameter when the relative improvement in the gradient is at
least as good as the improvement in the interior point parameter.
Otherwise, we do centering steps.  This doesn't appear, at the onset, to
work as well as just doing a constant reduction, so we keep the constant
decrease as default.  Finally, I largely removed most of the
trust-region scaling code.  I need to figure out what to do with this
code.  Honestly, it was always kind of irritating since it required
several additional checks in order to make sure everything worked
properly.  In addition, it appeared to really slow down the performance
of the algorithm.  The idea behind the scaling was to make the elements
near infeasibility very large so that the trust-region is violated
sooner.  In theory, that should keep us away from the boundary, but I
was never very impressed with the performance.  Really, if we want to
shape the trust-region, we should do so through the inner product as
that would be more natural through CG.  In any case, there's still some
remnants of the code that's not activated at the moment.