There's a problem with the orthogonality check...
There's a problem with the orthogonality check in CG when the
preconditioner is not a projection.  Basically, in preconditioned CG, we
have that <Br_i,r_j>=0 for i!=j.  So, we use this as a check to
determine when we start to lose orthogonality.  In order to do this, we
construct a matrix inv(D)*(BR)'*R*inv(D)-I where D=diag(||Br_i||) and
check its norm.  If we look at the digonal elements of this matrix, we
see that they are <B ri,ri>/||Bri|| ||Bri|| = <B ri,ri> / <Bri,Bri> = <B
ri, ri> / <B'B ri,ri> = <Bri,ri> / <Bri,ri> when we work in exact
arithmetic AND B is a projection.  When B is not a projection, this
reduction does not hold and even in exact arithmetic, the matrix
inv(D)*(BR)'*R*inv(D)-I is not zero.  This basically means that using a
preconditioner that is not the identity in the unconstrained code sets
off the orthongality check.  Until we come up with a better way of
resolving it, we've added a flag that turns off the orthogonality check
unless we're doing the inexact composite-step method.  --Miguel and Joe