I modified the SDP line search so that it...
I modified the SDP line search so that it correctly provides a
feasibility bound.  I'm still a little iffy with the Lanczos code that's
in there, so I've added an iteratively restarted Arnoldi method in
preparation for replacing this routine.  It's not currently turned on.
In addition, I added an example where we can read linear SDP problems in
SDPA format.  Mostly, this should provide an example of how the code can
be flexibily used.  At the moment, we can successfully solve some, but
not all of the problems.  Because we don't have a built-in factorization
routine, we don't have a good way to solve the augmented systems that
arise from these problems.  Hence, we instead use a purely inequality
constrained formulation to solve these problems.  This leads to a
stage-1/stage-2 process, which sort of works, but is a bit unusual for
these problems.  It may be just a matter of tuning the parameters, so we
need to investigate the behavior of this code further.