I updated the line search routine for SDP to...
I updated the line search routine for SDP to use an implicitely
restarted Arnoldi method.  It seems mostly reliable, but there are
instances where the method converges to the wrong eigenvalue.  Part of
the problem is that these Krylov methods tend to converge to the
smallest eigenvalue, but the eigenvalue estimate that they return only
tell us if we've converge to *some* eigenvalue.  If this is the wrong
one, the line search routine gets confused, we take a bad step, and the
algorithm fails.  In order to combat this, I've added a safe guard
procedure that checks whether or not X+alpha dX is positive definite.
This is done by taking a Choleski of the result and testing whether or
not that fails.  If it fails, we do a simple back tracking line search
until it suceeds.  In theory, since X is positive definite, there exists
some alpha so that this is true.