This is a fairly large checkin that overhauls...
This is a fairly large checkin that overhauls the interior point system.
Previously, we required the user to provide a Euclidean-Jordan algebra
when exercising algorithms that use inequality constraints.  For linear
and second-order cone programming, this works flawlessly.  However, for
semidefinite programming, this is more awkward.  Specifically, when we
define a EJA on SDP cones with the product x o y = (xy + yx)/2, we end
up having to solve the Sylvester equations for the operator linv, which
is expensive since it generally requires a complete Schur decomposition.
In order to eliminate this cost, I've relaxed the requirements on the
EJA so that we no longer require the product to be commutative.  I've
also added a new requirement that we have an operator called symm, which
stands for symmetrize, which modifies x so that L(symm(x)) is symmetric.
The reason for this change is that we can now define the product x o y =
xy and linv(x)=inv(x).  These operations, especially the later, are
vastly cheaper to compute.  In order to allow this, though, I needed to
rederive the interior point equations so that we no longer use symmetry
in the reductions.  In the end, this actually simplifies our Hessian
modification so that we do slightly less computation and have a cleaner
result than before.  Before, we had to do a hand symmetrization so that
the Hessian was self-adjoint.  In any case, this algebra is not exactly
a Euclidean-Jordan algebra, but close.  The big requirement now is that
we have to be careful on how we define our inequality constraint, h.
Really, we need h(x) to be symmetric, but produce a result in a space
that can handle nonsymmetry.  For example, for matrices, we need the
space Z to be all m x m matrices, even no symmetric ones, but we need
h(x) to produce a symmetric matrix.  This has impact on the operator
h'(x)*.  Previously, I was really assuming that the element being passed
into this operator was symmetric, but we can't assume that now.  If we
mess this up, we'll see that we're getting a larger number of Krylov
iterations in truncated-CG than we should.  This is because the Hessian
is nonsymmetric.  Perhaps we should add a test for symmetry in the
future.  By the way, it's technically possible to enforce symmetry
internally with no input from the user, but I couldn't find a good way
to make it numerically stable.  Basically, I would symmetrize the
equations for dz.  Unfortunately, small amounts of roundoff error would
creep into this computation and eventually the dual variable would
become nonsymmetric.  In any case, this new scheme appears to work much
better than the old.  It is also backward compatible with the old way of
defining the EJA as long as the symm operator is present and does
nothing.  In addition, I removed all of the forced casts to Integer and
Natural and instead allow C++ to automatically cast these values.
Mostly, the forced casting is useful if we were to allow the user to
define the Natural and Integer types.  However, since we use these
elements to index things like std::vectors, they must always be valid
C++ integers, signed or otherwise.  As such, we're restricting
everything so that our code is the only code that defines Integer and
Natural, which is how things are set up now.  If this changes in the
future, we'll have to reconsider these casts by hand.  The rest of the
changes to the unit tests are done in response to the new interior point
code.  They don't change that much, but the parameters needed to be
tuned slightly.