I updated the example where we solve linear...
I updated the example where we solve linear SDP problems stored in the
sparse sdpa format.  At this point, we can solve most of the problems in
SDPLIB, which should be a good enough verification that our SDP solver
is working correctly.  Unfortunately, the parameters in order to make
these solves works varies from problem to problem.  I'm optimistic that
using the equality constrained formulation of this problem will fix
this, but I'll have to try later.  Mostly, the equality constrained
formulations require a preconditioner and I've not yet written one.
Until that time, we cheat and use a purely inequality constrained
formulation in order to solve the problems.  This is a little bit funny
because we're using a primal dual algorithm to obtain a solution that
many people would consider to be a pure dual solution.  In any case, all
of these examples can be improved upon and this was meant just to be a
demonstration of capability.