I forgot to check in the updated CMakeLists.
I forgot to check in the updated CMakeLists.txt for the last example.
In addition, I'm adding a different example that shows one possible
strategy for finding a feasible solution to an inequality constrained
problem without adding an entire slack.  Honestly, I think the interior
point algorithm has trouble on this formulation, but I've not explored
it far enough.  It provides a contrast to the phase-1 phase-2 approach
used in the sparse sdpa example.