I fixed some bugs in peopt where we were using...
I fixed some bugs in peopt where we were using the wrong vector space in
an operation.  Really, the way to find these bugs is to make sure that
we use three different, incompatible vector spaces in our unit tests.
Then, the C++ compiler should find these errors due to templating.  In
addition, I've done a little bit of work in GMRES, but decided to not do
some fixes at this point in time.  Specifically, GMRES is a little bit
tricky when working with general vector spaces since it's hard to
determine when we have breakdown.  For example, if we are using GMRES on
a 5x5 system, we're guaranteed to get breakdown at the 5th iteration at
the latest.  Now, for really poorly conditioned systems and with the
dyanmic checks that the inexact composite step SQP algorithm have, it's
hard to detect this.  Then, we end up added an extra vector and our
solves go all wonky.  In order to detect this, I added code where we do
the triangular solve with R after the QR factorization.  Here, most of
the time, R is well-conditioned.  If all of the sudden R becomes
singular, it's probably due to some sort of breakdown.  Now, this code
by itself works fine, but I was using routines out of LAPACK in order to
calculate the reciprocal of the condition number.  This added a
dependency to LAPACK for the core optimization routines and made it so
that we can't be deployed in a header only manner.  As such, until I
duplicate the code in LAPACK in order to calculate this condition
number, the code is ifdef'd out.  Later on, I need to reactive this
code.