I improved the error handling so that, in...
I improved the error handling so that, in theory, we should not segfault
if the user has an error in one of the functions provided to peopt.  I
had some trouble getting this to work on R2010a, but it seems to work
fine on R2013a, so we may have some versioning issues.  In addition,
it's difficult sometimes to diagnose where the error is given the output
from these functions.  For example, if the gradient gives an erroneous
answer such as a vector that's too large, it's hard to trap that in the
gradient function call itself.  Likely, the first time we'll see this
error is going to be in something like the innr function called from
inside truncated CG.  I don't really know how to improve this situation
since we don't know if our data is good or not.  However, in theory, the
user should be able to figure out the casuse of the error from the
Matlab debugger and some intution.  Nonetheless, it's worthwhile to
figure out a better way to do this.