I did another reworking of the interior point...
I did another reworking of the interior point code in addition to some
more minor bug fixes.  First, it turns out that we were reporting the
relative error in CG wrong if we were using a preconditioner.  When
calculating the relative error, we need to normalize against the
preconditioned right hand side, otherwise it looks like the error went
up.  Second, I fixed the caching for the equality and inequality
modification code so that they use the exact same code for calculating
the relative error when caching data.  This code has been moved into the
linear algebra.  Third, I modified how we calculate our initial
Langrange multiplier for the inequality constraints.  Previously, we
chose a multiple of the identity element that had norm equal to the norm
of the inequality constraint.  Instead, we now choose an initial
multiplier to the a multiple of the identity so that <h(x),z>/m = 1.
This makes is much, much easier to choose the parameter eps_mu for the
interior point stopping condition.  In addition, it seems to work a bit
better in situations when h(x) is very large.  Fourth, I reworked the
sparse SDPA example code to now use an exact preconditioner for its
solves.  Well, it mostly uses an exact preconditioner until the
condition number of the preconditioner reaches some large value.  In any
case, this appears to have made the example code work more reliably on
the sdplib problems without having to hand tune the paramters nearly as
much.  It's still imperfect and if we look at the conditioner number of
the precondioner as well as the Hessian, they're pretty bad for many of
these problems.  What I can't quite figure out is why we sometimes gets
stagnation on a problem even though the solution looks good.  Perhaps
all we need is a minor fix to the optimization parameters, but I suspect
that the Lagrange multiplier is a little off.  Since we have basically
the right solution, it seems like the primal variable should be ok,
which leaves the dual variables.  In any case, I removed the directory
with the individual problem setup files for problems in sdplib.
Technically, the solver would work better if we updated these files for
use with the new interior point code described above.  However, at this
point, the problems mostly work and we can tune the behavior of these
problems further later.