I modified how the actual and predicted...
I modified how the actual and predicted reduction are reported.
Previously, the output contained the ratio ared/pred.  Now, it contains
ared and pred separately.  This is similar to a problem and fix that I
made some time ago where the two variables were separated.  Basically,
lots of bad things can happen in an optimization and it's hard to print
this out in a simple ratio.  For example, in theory, pred should always
be positive.  Sometimes, especially if we have a nonsymmetric Hessian,
this is negative.  Now, sometimes, ared or pred could be a NaN.  This
happens when a user's function is evaluated at a point that causes an
issue.  Anyway, the way the output worked previously was that we
reported the ratio ared/pred, but if pred was negative, we may it NaN.
This hid any actual NaNs that were thrown, which is kind of irritating.
Returning the raw ratio wasn't a good idea because if ared and pred were
negative, it would look like we rejected a perfectly good step.  In
order to completely avoid these issues, we now report the raw ared and
pred, which should contain complete information.