I modified how the trust-region was expanded...
I modified how the trust-region was expanded in the reduced-space code.
Previously, we only expanded the trust-region if the step forced us into
the trust-region and the ratio between the actual and predicted
reduction was high.  Generally, this worked pretty well, but there are a
couple of cases where this is inconvenient.  For example, in a
trust-region method, we may hit the edge of the trust-region, then cut
back the step in order to satisfy feasibility.  In this case, increasing
the size of the trust-region may be the right thing to do even though
the final iterate was cut back since doing more work on the trust-region
subproblem means that we may have would have had a better step to cut
back.  Another place where this comes up is when implementing a
projected gradient method using the state manipulator.  Even though we
cut back a step, it may be advantageous to increase the size of the
trust-region.  So, in order to figure this out, we just look at why the
truncated Krylov method exited.  If it was due to hitting the
trust-region or negative curvature, we know that we should expand the
size of the trust-region.  Funny enough, I already added this code to
the equality constrained section, so this puts the unconstrained code in
line with that.