I modified the inequality constrained algebra...
I modified the inequality constrained algebra for srch so that srch <-
argmax {alpha \in Real >= 0 : alpha x + y >= 0} where y >0.  Previously,
if this number was infinity, we returned -1.  Now, we actually return an
inf.  Basically, this makes the code more consistent with the
mathematical definition.  In addition, it actually simplifies a bit of
the line-search code in the vector spaces since x < infinity is true
when x is finite.  The drawback to this approach is that we're requiring
IEEE compliant arithmetic, which some users don't like to do since
aggressive optimizations break this.  However, we're already pretty
dependent on treating NaNs properly, so this really isn't a bigger
burden.