I put in a couple of fixes.
I put in a couple of fixes.  First, in the interior point method for
SDP, there's a safeguard where we take the Choleski factorization of the
new iterate to insure it's positive definite.  As it turns out,
sometimes this check passes, but the iterate is indefinite.  In other
words, depending on the round-off error, a Choleski factorization may
exist for an indefinite matrix.  I believe the book 'Accuracy and
Stability of Numerical Algorithms' by Higham has details.  Anyway, this
caused the code to hit an infinite loop eventually in the srch function
since the previous iterate was indefinite and then it really can't find
a step to become positive definite.  In order to get around this, I have
a temporary fix where if the new line-search parameter becomes 0, we
bail.  I don't really like this fix since it could take awhile for this
to occur, so we need a better long term solution later. Next, I finished
eliminating the last remnants of delta_max.  In all honesty, this
parameter has caused me no end of trouble working with other codes.
Basically, no one ever sets it and it ends up cutting off the steps for
poorly scaled problems.  Since the code has modified the trust-region
radius based on need for some time, this code was largely irrelavant
anyway.  Meaning, a traditional trust-region method will increase the
size of the trust-region radius when the actual versus predicted
reduction is above some threshold.  For us, by default, it's 0.9.  The
standard algorithm does this regardless if the current iterate is near
the trust-region.  In this code, I don't do that.  Instead, I only
increase the size of the trust-region radius if the iterate is limited
by the trust-region radius.  I've done some refinement of this idea over
time.  Originally, I did a hard check to see if we were close using
norms.  Recently, I changed this to look at why the Krylov method
stopped and then increase the trust-region method if the method hit the
boundary.  This seemed to work better since we may cut the step length
after the fact due to the interior point method or a projected gradient
algorithm.  Anyway, since we don't increase the size of the trust-region
unless it's necessary, we don't have to worry about the trust-region
growing too an unmanagable size, which means we don't really need
delta_max.  Now, I had to change the ScaledIdentity operator to cope
with the change, but now it works fine.  Basically, I force the
algorithm to take a step at two times the trust-region, which forces us
into the trust-region, which should recreate the trust-region equivalent
of steepest descent for poorly scaled problems.  In addition, I had to
rework quite a few of the unit tests.  This is because the new default
starting delta is now 1 instead of 100.  It turns out that cutting off
the interior point method early in the optimization can cause issues
since the interior point paramter drops too fast and the algorithm gets
behind.  Maybe there's a larger issue at play, which should be
investigated.  Until that time, note that the starting step size can
have an impact on the interior point performance.