So it turns out that I'm terrible at dividing...
So it turns out that I'm terrible at dividing floating point numbers in
my head.  As such, I've reintroduced the ared/pred output.  At the same
time, I'm keeping ared and pred separate in order to diagnose a variety
of issues that I discussed when I separated the output in the first
place.  Also, I added a small bit of code to stop instantiation of the
Printer class in case for some odd reason a user, or myself for that
matter, try to.  It really is part of this templated namespace trick I
use, so it should not be instantiated.